requirements.txt
asgiref==3.7.2
astroid==3.0.1
dill==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
PyJWT==2.8.0
pylint==3.0.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing==3.7.4.3
typing_extensions==4.8.0
